DELFT3D-37357: Solve De;ft3D-FLOW testbank issues Problems with plugin modules solved (compare with revision 6619, 30 September 2016):
rdcul.f90, rdtrafrm.f90, rdsed.f90:
  When writing in string "rec", do not use "rec" as input

shared_library_fortran_api.c:
  The DLL pointer must be of type "long long int"

culver.f90, erosilt.f90, eqsettle.f90:
  Added: message_c: c-copy of parameter message: copy message to message_c, pass message_c into "PERFORM_FUNCTION_*", copy message_c to message

f902c.c, shared_lib_fallve.c, shared_lib_trafrm.c:
  "PERFORM_FUNCTION_*": first argument (DLLHandle) has type "long long int", last parameter (length_message) must be removed because it is now a c-string

plugin_culvert.f90, settling_velocity.f90, partheniades_krone.f90:
  error_message_c is passed through via the parameter list. When error_message is changed, copy it into error_message_c

+++ src/utils_lgpl/deltares_common/packages/deltares_common_c/src/shared_library_fortran_api.c
@@ -111,7 +111,7 @@ void RemoveTrailingBlanks_dll(char * String)
  * ============================================================================
  */
 #if defined(WIN32) || defined (HAVE_CONFIG_H)
-long STDCALL OPEN_SHARED_LIBRARY(int64_t * sharedDLLHandle, char * library, long length_lib)
+long STDCALL OPEN_SHARED_LIBRARY(long long int * sharedDLLHandle, char * library, long length_lib)
 #elif defined (salford32)
 extern "C" OPEN_SHARED_LIBRARY(int64_t * sharedDLLHandle, char * library, long length_lib)
 #endif
@@ -136,7 +136,7 @@ extern "C" OPEN_SHARED_LIBRARY(int64_t * sharedDLLHandle, char * library, long l
     if (tmpSharedDLL->dllHandle != NULL)
     {
         error = 0;
-        *sharedDLLHandle = (int64_t) tmpSharedDLL;
+        *sharedDLLHandle = (long long int) tmpSharedDLL;
     }
 
     free(lib_name); lib_name = NULL;
